Cast the gathered lm_head weight once instead of once per chunk
Correcting the previous commit: measuring the SFT path showed both operands arriving as bf16, and that
does not generalise. SFT projects inside the patched `lm_head.forward`, where FSDP2 has already gathered
and cast the parameter. The divergence loss projects outside any forward, so it sees the fp32 master
weight: probed on a 2-GPU FSDP2 bf16 run, `lm_head.weight` there is a DTensor of fp32 and `full_tensor()`
returns fp32. Nothing casts it back either, since accelerate's autocast wraps the model's forward and
`Trainer.autocast_smart_context_manager` is a null context.

So the cast in `_chunk` does real work here, and doing it per chunk rebuilt a vocab-sized copy every
time, plus once more per gradient-checkpoint recompute. It now rides with the `full_tensor()` call that
is already hoisted out of the loop. The per-chunk cast stays for the ZeRO-3 path, where the weight is
only gathered inside `maybe_gather_lm_head_ctx`.

The test covers both projections, the student's and the teacher's.

+++ trl/trainer/distillation_trainer.py
@@ -232,12 +232,14 @@ def _chunked_divergence_loss(
     # gradient-checkpointed chunk loop causes FSDP2 to re-gather it once per chunk during backward recomputation.
     # full_tensor() converts it to a plain tensor once; all chunks reference that tensor, so only one all-gather occurs
     # (in full_tensor()'s backward). Done per model since the student and teacher have their own heads.
+    # Under mixed precision the gathered weight is the fp32 master copy, so cast it to the hidden-states dtype here:
+    # the projection needs it in that dtype, and doing it once beats rebuilding a vocab-sized copy per chunk.
     if isinstance(student_lm_head_weight, torch.distributed.tensor.DTensor):
-        student_lm_head_weight = student_lm_head_weight.full_tensor()
+        student_lm_head_weight = student_lm_head_weight.full_tensor().to(student_hidden_states.dtype)
         if student_lm_head_bias is not None:
             student_lm_head_bias = student_lm_head_bias.full_tensor()
     if isinstance(teacher_lm_head_weight, torch.distributed.tensor.DTensor):
-        teacher_lm_head_weight = teacher_lm_head_weight.full_tensor()
+        teacher_lm_head_weight = teacher_lm_head_weight.full_tensor().to(teacher_hidden_states.dtype)
         if teacher_lm_head_bias is not None:
             teacher_lm_head_bias = teacher_lm_head_bias.full_tensor()
